video: rockchip/mpp_device: new common device design
--------------    ---------------    ------------------
|            |    |             |    | mpp device     |
| mpp device |<-->| mpp service |<-->|  implementation|
|    top     |    |             |    |  and buttom    |
--------------    ---------------    ------------------

The mpp common device will do the the top part, but still
allow the mpp device implementation change the default behaviour.

The mpp device implementation will register itself to the mpp
device framework. But the mpp common device is the base
class not the top class. The  mpp common device framework
also in charge of the communication with mpp service.

The clocks and power are all moved into the runtime power
domain management, we don't need to assign the clocks
in the driver but leaves those work to the rockchip power
domain and common driver, we just need to configure the clocks
in our device node.

The link mode will be regard as a bundle of tasks.

The functions with rockchip_mpp_* prefix are used
internal for the task operation.

This device mode is not just suitable for the video IP but
also IEP I think.

I now make the mpp common device become a module and
a different mpp device implementation will be module as well.

The iommu and video acceleration can work independently.

Change-Id: I97dd4eaf20a925f820133a6cb2d954fb97367cb3
Signed-off-by: Randy Li <randy.li@rock-chips.com>
Signed-off-by: ayaka <ayaka@soulik.info>